requirements.txt
black==24.8.0
click==8.1.7
mypy-extensions==1.0.0
packaging==24.1
paho-mqtt==2.1.0
pathspec==0.12.1
platformdirs==4.2.2
